+++ TRNG-Pendel/OldCode/pendulumPicToBin.py
